+++ lib/awful/_compat.lua
@@ -1,0 +1,27 @@
+-- This file contains all global backward compatibility workarounds for the
+-- Core API changes.
+local gtimer = require("gears.timer")
+local util   = require("awful.util" )
+local spawn  = require("awful.spawn")
+local gdebug = require("gears.debug")
+
+function timer(...) -- luacheck: ignore
+    gdebug.deprecate("gears.timer", {deprecated_in=4})
+    return gtimer(...)
+end
+
+util.spawn = function(...)
+   gdebug.deprecate("awful.spawn", {deprecated_in=4})
+   return spawn.spawn(...)
+end
+
+util.spawn_with_shell = function(...)
+   gdebug.deprecate("awful.spawn.with_shell", {deprecated_in=4})
+   return spawn.with_shell(...)
+end
+
+util.pread = function()
+    gdebug.deprecate("Use io.popen() directly or look at awful.spawn.easy_async() "
+            .. "for an asynchronous alternative", {deprecated_in=4})
+    return ""
+end
